\name{aberrant}
\alias{aberrant}

\title{
Outlier identification
}

\description{
Outlier identification based on two summary statistics.
}

\usage{
aberrant(x, lambda, niter = 10000, burnin = 100, prior_df = NULL, prior_scale = NULL, hyper_prior_mean = NULL, hyper_prior_var = NULL, hyper_prior_df = NULL, hyper_prior_scale = NULL, alpha = NULL, beta = NULL, standardize = TRUE, verbose = TRUE, uncorr = FALSE)
}

\arguments{
  \item{x}{
Vector, matrix or numeric data frame, with two summary statistics in column.
}
  \item{lambda}{
Ratio of the standard deviations of outlying and normal individuals.
}
  \item{niter}{
Number of samples to be generated by the Gibbs sampling. Default to 10000.
}
  \item{burnin}{
Number of burn-in samples. Default to 100.
}
  \item{prior_df}{
Degree(s) of freedom of the distribution that describes prior information on the covariance matrix of the summary statistics for the normal individuals. If uncorr=FALSE, degree of freedom of length 1 of an Inverse-Wishart distribution. If uncorr=TRUE, vector of degrees of freedom of length 2 for the 2 Scaled Inverse Chi-Square priors.
}
  \item{prior_scale}{
Scale matrix of the distribution that describes prior information on the covariance matrix of the summary statistics for the normal individuals. If uncorr=FALSE, scale matrix of an Inverse-Wishart distribution. If uncorr=TRUE, diagonal scale matrix describing scale parameters of the 2 Scaled Inverse Chi-Square priors.
}
  \item{hyper_prior_mean}{
Means of the normal priors for the mean hyper-parameters. Vector of length 2.
}
  \item{hyper_prior_var}{
Variances of the normal priors for the mean hyper-parameters. Vector of length 2.
}
  \item{hyper_prior_df}{
Degrees of freedom of the Scaled Inverse Chi-Square priors for the variance hyper-parameters. Vector of length 2.
}
  \item{hyper_prior_scale}{
Scale parameters of the Scaled Inverse Chi-Square priors for the variance hyper-parameters. Vector of length 2.
}
  \item{alpha}{
First shape parameter of the Beta distribution describing prior information on the probability that an individual is an outlier.
}
  \item{beta}{
Second shape parameter of the Beta distribution describing prior information on the probability that an individual is an outlier.
}
  \item{standardize}{
A logical indicating whether the summary statistics should be standardized.
}
  \item{verbose}{
logical. If TRUE, verbose output is generated during the Gibbs sampling. 
}
  \item{uncorr}{
logical. If TRUE, summary statistics are considered independent.
}
}

\details{
Prior parameters  'prior_df', 'prior_scale', 'hyper_prior_mean', 'hyper_prior_var', 'hyper_prior_df' and 'hyper_prior_scale' must be all NULL or all numeric. Prior parameters 'alpha' and 'beta' must be both NULL or both numeric. If prior parameters are not specified, then uninformative priors are used.
}

\value{
\item{x }{Initial data matrix with the 2 summary statistics.}
\item{group }{Vector indicating whether an individual is an outlier (=1) or not (=0). Individuals are in the same order as in the initial data matrix x.}
\item{posterior }{Vector indicating the posterior probability for an individual to be an outlier. }
\item{lambda }{Ratio of the standard deviations of outlying and normal individuals used. }
\item{post_mean }{Posterior mean of the summary statistics for the normal individuals. }
\item{post_var }{Posterior covariance matrix of the summary statistics for the normal individuals. }
\item{standardize }{Logical indicating if summary statistics were standardized. }
\item{inlier }{Indices of normal individuals. }
\item{outlier }{Indices of outlying individuals. }
}

\references{
Celine Bellenguez, Amy Strange, Colin Freeman, Wellcome Trust Case Control Consortium 2, Chris CA Spencer. A robust clustering algorithm for identifying problematic samples in genome-wide association studies. Bioinformatics.
}
\author{
Celine Bellenguez and Chris CA Spencer
}

\examples{
x<-rmvt(1000, sigma=matrix ( c ( 1, 0.5 , 0.5 , 1 ) , 2 , 2 ), df=3)
aberrant (x, lambda=30, alpha=1, beta=20)
}